requirements.txt
stripe>=5.0.0
django-cors-headers>=4.0.0
